moving things around, adding base analysis class, started clusterer
